Split out types for easier tests
