+++ edge-hub/src/Microsoft.Azure.Devices.Edge.Hub.Core/twin/StoringTwinManager.cs
@@ -6,14 +6,12 @@ namespace Microsoft.Azure.Devices.Edge.Hub.Core.Twin
     using System.Collections.Concurrent;
     using System.Collections.Generic;
     using System.Threading.Tasks;
-    using Microsoft.Azure.Devices.Edge.Hub.Core.Cloud;
     using Microsoft.Azure.Devices.Edge.Hub.Core.Device;
     using Microsoft.Azure.Devices.Edge.Hub.Core.Identity;
     using Microsoft.Azure.Devices.Edge.Storage;
     using Microsoft.Azure.Devices.Edge.Util;
     using Microsoft.Azure.Devices.Shared;
     using Microsoft.Extensions.Logging;
-    using Nito.AsyncEx;
 
     public class StoringTwinManager : ITwinManager
     {
@@ -21,27 +19,29 @@ namespace Microsoft.Azure.Devices.Edge.Hub.Core.Twin
         readonly IMessageConverter<Twin> twinConverter;
         readonly IConnectionManager connectionManager;
         readonly IValidator<TwinCollection> reportedPropertiesValidator;
-        readonly ReportedPropertiesStore reportedPropertiesStore;
-        readonly TwinStore twinEntityStore;
-        readonly CloudSync cloudSync;
+        readonly IReportedPropertiesStore reportedPropertiesStore;
+        readonly ITwinStore twinStore;
+        readonly ICloudSync cloudSync;
         readonly TimeSpan twinSyncPeriod;
         readonly ConcurrentDictionary<string, DateTime> twinSyncTime = new ConcurrentDictionary<string, DateTime>();
 
-        StoringTwinManager(
+        internal StoringTwinManager(
             IConnectionManager connectionManager,
             IMessageConverter<TwinCollection> twinCollectionConverter,
             IMessageConverter<Twin> twinConverter,
             IValidator<TwinCollection> reportedPropertiesValidator,
-            IEntityStore<string, TwinStoreEntity> twinStore,
+            ITwinStore twinStore,
+            IReportedPropertiesStore reportedPropertiesStore,
+            ICloudSync cloudSync,
             TimeSpan twinSyncPeriod)
         {
             Preconditions.CheckNotNull(twinStore, nameof(twinStore));
             this.connectionManager = Preconditions.CheckNotNull(connectionManager, nameof(connectionManager));
             this.twinCollectionConverter = Preconditions.CheckNotNull(twinCollectionConverter, nameof(twinCollectionConverter));
             this.twinConverter = Preconditions.CheckNotNull(twinConverter, nameof(twinConverter));
-            this.cloudSync = new CloudSync(connectionManager, twinCollectionConverter, twinConverter);
-            this.twinEntityStore = new TwinStore(twinStore);
-            this.reportedPropertiesStore = new ReportedPropertiesStore(twinStore, this.cloudSync);
+            this.cloudSync = Preconditions.CheckNotNull(cloudSync, nameof(cloudSync));
+            this.twinStore = Preconditions.CheckNotNull(twinStore, nameof(twinStore));
+            this.reportedPropertiesStore = Preconditions.CheckNotNull(reportedPropertiesStore, nameof(reportedPropertiesStore));
             this.reportedPropertiesValidator = reportedPropertiesValidator;
             this.twinSyncPeriod = twinSyncPeriod;
         }
@@ -49,23 +49,28 @@ namespace Microsoft.Azure.Devices.Edge.Hub.Core.Twin
         public static ITwinManager Create(
             IConnectionManager connectionManager,
             IMessageConverterProvider messageConverterProvider,
-            IEntityStore<string, TwinStoreEntity> twinStore,
+            IEntityStore<string, TwinStoreEntity> entityStore,
             IDeviceConnectivityManager deviceConnectivityManager,
             IValidator<TwinCollection> reportedPropertiesValidator,
             TimeSpan twinSyncPeriod)
         {
             Preconditions.CheckNotNull(connectionManager, nameof(connectionManager));
             Preconditions.CheckNotNull(messageConverterProvider, nameof(messageConverterProvider));
-            Preconditions.CheckNotNull(twinStore, nameof(twinStore));
+            Preconditions.CheckNotNull(entityStore, nameof(entityStore));
             Preconditions.CheckNotNull(deviceConnectivityManager, nameof(deviceConnectivityManager));
             Preconditions.CheckNotNull(reportedPropertiesValidator, nameof(reportedPropertiesValidator));
 
+            IMessageConverter<TwinCollection> twinCollectionConverter = messageConverterProvider.Get<TwinCollection>();
+            IMessageConverter<Twin> twinConverter = messageConverterProvider.Get<Twin>();
+            ICloudSync cloudSync = new CloudSync(connectionManager, twinCollectionConverter, twinConverter);
             var twinManager = new StoringTwinManager(
                 connectionManager,
-                messageConverterProvider.Get<TwinCollection>(),
-                messageConverterProvider.Get<Twin>(),
+                twinCollectionConverter,
+                twinConverter,
                 reportedPropertiesValidator,
-                twinStore,
+                new TwinStore(entityStore),
+                new ReportedPropertiesStore(entityStore, cloudSync),
+                cloudSync,
                 twinSyncPeriod);
 
             deviceConnectivityManager.DeviceConnected += (_, __) => twinManager.DeviceConnectedCallback();
@@ -101,7 +106,7 @@ namespace Microsoft.Azure.Devices.Edge.Hub.Core.Twin
 
         async Task SyncTwinAndSendDesiredPropertyUpdates(string id)
         {
-            Option<Twin> storeTwin = await this.twinEntityStore.Get(id);
+            Option<Twin> storeTwin = await this.twinStore.Get(id);
             if (!storeTwin.HasValue)
             {
                 Events.NoSyncTwinNotStored(id);
@@ -119,7 +124,7 @@ namespace Microsoft.Azure.Devices.Edge.Hub.Core.Twin
                                 async cloudTwin =>
                                 {
                                     Events.UpdatingTwinOnDeviceConnect(id);
-                                    await this.twinEntityStore.Update(id, cloudTwin);
+                                    await this.twinStore.Update(id, cloudTwin);
                                     this.twinSyncTime.AddOrUpdate(id, DateTime.UtcNow, (_, __) => DateTime.UtcNow);
 
                                     string diffPatch = JsonEx.Diff(twin.Properties.Desired, cloudTwin.Properties.Desired);
@@ -142,11 +147,12 @@ namespace Microsoft.Azure.Devices.Edge.Hub.Core.Twin
         public async Task<IMessage> GetTwinAsync(string id)
         {
             Option<Twin> twinOption = await this.cloudSync.GetTwin(id);
-            Twin twin = await twinOption.Map(
+            Twin twin = await twinOption
+                .Map(
                     async t =>
                     {
                         Events.GotTwinFromCloud(id);
-                        await this.twinEntityStore.Update(id, t);
+                        await this.twinStore.Update(id, t);
                         this.twinSyncTime.AddOrUpdate(id, DateTime.UtcNow, (_, __) => DateTime.UtcNow);
                         return t;
                     })
@@ -154,7 +160,7 @@ namespace Microsoft.Azure.Devices.Edge.Hub.Core.Twin
                     async () =>
                     {
                         Events.GettingTwinFromStore(id);
-                        Option<Twin> storedTwin = await this.twinEntityStore.Get(id);
+                        Option<Twin> storedTwin = await this.twinStore.Get(id);
                         return storedTwin.GetOrElse(() => new Twin());
                     });
             return this.twinConverter.ToMessage(twin);
@@ -164,7 +170,7 @@ namespace Microsoft.Azure.Devices.Edge.Hub.Core.Twin
         {
             Events.UpdatingDesiredProperties(id);
             TwinCollection patch = this.twinCollectionConverter.FromMessage(twinCollection);
-            await this.twinEntityStore.UpdateDesiredProperties(id, patch);
+            await this.twinStore.UpdateDesiredProperties(id, patch);
             await this.SendPatchToDevice(id, twinCollection);
         }
 
@@ -173,7 +179,7 @@ namespace Microsoft.Azure.Devices.Edge.Hub.Core.Twin
             Events.UpdatingReportedProperties(id);
             TwinCollection patch = this.twinCollectionConverter.FromMessage(twinCollection);
             this.reportedPropertiesValidator.Validate(patch);
-            await this.twinEntityStore.UpdateReportedProperties(id, patch);
+            await this.twinStore.UpdateReportedProperties(id, patch);
             await this.reportedPropertiesStore.Update(id, patch);
             this.reportedPropertiesStore.InitSyncToCloud(id);
         }
@@ -200,263 +206,6 @@ namespace Microsoft.Azure.Devices.Edge.Hub.Core.Twin
                 });
         }
 
-        internal class ReportedPropertiesStore
-        {
-            readonly IEntityStore<string, TwinStoreEntity> twinStore;
-            readonly CloudSync cloudSync;
-            readonly AsyncLockProvider<string> lockProvider = new AsyncLockProvider<string>(10);
-            readonly AsyncAutoResetEvent syncToCloudSignal = new AsyncAutoResetEvent(false);
-            readonly ConcurrentQueue<string> syncToCloudQueue = new ConcurrentQueue<string>();
-            readonly Task syncToCloudTask;
-
-            public ReportedPropertiesStore(IEntityStore<string, TwinStoreEntity> twinStore, CloudSync cloudSync)
-            {
-                this.twinStore = twinStore;
-                this.cloudSync = cloudSync;
-                this.syncToCloudTask = this.SyncToCloud();
-            }
-
-            public async Task Update(string id, TwinCollection patch)
-            {
-                using (await this.lockProvider.GetLock(id).LockAsync())
-                {
-                    Events.StoringReportedPropertiesInStore(id, patch);
-                    await this.twinStore.PutOrUpdate(
-                        id,
-                        new TwinStoreEntity(patch),
-                        twinInfo =>
-                        {
-                            Events.UpdatingReportedPropertiesInStore(id, patch);
-                            TwinCollection updatedReportedProperties = twinInfo.ReportedPropertiesPatch
-                                .Map(reportedProperties => new TwinCollection(JsonEx.Merge(reportedProperties, patch, /*treatNullAsDelete*/ false)))
-                                .GetOrElse(() => patch);
-                            return new TwinStoreEntity(twinInfo.Twin, Option.Maybe(updatedReportedProperties));
-                        });
-                }
-            }
-
-            public void InitSyncToCloud(string id)
-            {
-                this.syncToCloudQueue.Enqueue(id);
-                this.syncToCloudSignal.Set();
-            }
-
-            async Task SyncToCloud()
-            {
-                while (true)
-                {
-                    try
-                    {
-                        await this.syncToCloudSignal.WaitAsync();
-                        while (this.syncToCloudQueue.TryDequeue(out string id))
-                        {
-                            await this.SyncToCloud(id);
-                        }
-                    }
-                    catch (Exception e)
-                    {
-                        Events.ErrorSyncingReportedPropertiesToCloud(e);
-                    }
-                }
-            }
-
-            public async Task SyncToCloud(string id)
-            {
-                Events.SyncingReportedPropertiesToCloud(id);
-                Option<TwinStoreEntity> twinWithReportedProperties = (await this.twinStore.Get(id))
-                    .Filter(t => t.ReportedPropertiesPatch.HasValue);
-                if (twinWithReportedProperties.HasValue)
-                {
-                    using (await this.lockProvider.GetLock(id).LockAsync())
-                    {
-                        Events.StoredReportedPropertiesFound(id);
-                        Option<TwinStoreEntity> twinInfo = await this.twinStore.Get(id);
-                        await twinInfo.ForEachAsync(
-                            async ti =>
-                            {
-                                await ti.ReportedPropertiesPatch.ForEachAsync(
-                                    async reportedPropertiesPatch =>
-                                    {
-                                        bool result = await this.cloudSync.UpdateReportedProperties(id, reportedPropertiesPatch);
-
-                                        if (result)
-                                        {
-                                            Events.UpdateReportedPropertiesSucceeded(id);
-                                            await this.twinStore.Update(
-                                                id,
-                                                t => new TwinStoreEntity(t.Twin, Option.None<TwinCollection>()));
-                                            twinInfo = await this.twinStore.Get(id);
-                                        }
-                                        else
-                                        {
-                                            Events.UpdateReportedPropertiesFailed(id);
-                                        }
-                                    });
-                            });
-                    }
-                }
-                else
-                {
-                    Events.DoneSyncingReportedProperties(id);
-                }
-            }
-        }
-
-        internal class TwinStore
-        {
-            readonly IEntityStore<string, TwinStoreEntity> twinStore;
-
-            public TwinStore(IEntityStore<string, TwinStoreEntity> twinStore)
-            {
-                this.twinStore = twinStore;
-            }
-
-            public async Task<Option<Twin>> Get(string id)
-            {
-                Option<TwinStoreEntity> twinStoreEntity = await this.twinStore.Get(id);
-                return twinStoreEntity.FlatMap(t => t.Twin);
-            }
-
-            public async Task UpdateReportedProperties(string id, TwinCollection patch)
-            {
-                Events.UpdatingReportedProperties(id);
-                await this.twinStore.PutOrUpdate(
-                    id,
-                    new TwinStoreEntity(patch), 
-                    twinInfo =>
-                    {
-                        twinInfo.Twin
-                            .ForEach(
-                                twin =>
-                                {
-                                    TwinProperties twinProperties = twin.Properties ?? new TwinProperties();
-                                    TwinCollection reportedProperties = twinProperties.Reported ?? new TwinCollection();
-                                    string mergedReportedPropertiesString = JsonEx.Merge(reportedProperties, patch, /* treatNullAsDelete */ true);
-                                    twinProperties.Reported = new TwinCollection(mergedReportedPropertiesString);
-                                    twin.Properties = twinProperties;
-                                    Events.MergedReportedProperties(id);
-                                });
-                        return twinInfo;
-                    });
-            }
-
-            public async Task UpdateDesiredProperties(string id, TwinCollection patch)
-            {
-                Events.UpdatingDesiredProperties(id);
-                Option<Twin> storedTwin = await this.Get(id);
-                if(storedTwin.HasValue)
-                {                await this.twinStore.Update(
-                    id,
-                    twinInfo =>
-                    {
-                        twinInfo.Twin
-                            .ForEach(
-                                twin =>
-                                {
-                                    TwinProperties twinProperties = twin.Properties ?? new TwinProperties();
-                                    TwinCollection desiredProperties = twinProperties.Desired ?? new TwinCollection();
-                                    if (desiredProperties.Version + 1 == patch.Version)
-                                    {
-                                        string mergedDesiredPropertiesString = JsonEx.Merge(desiredProperties, patch, /* treatNullAsDelete */ true);
-                                        twinProperties.Desired = new TwinCollection(mergedDesiredPropertiesString);
-                                        twin.Properties = twinProperties;
-                                        Events.MergedDesiredProperties(id);
-                                    }
-                                });
-
-                        return twinInfo;
-                    });
-                }
-                else
-                {
-                    Events.NoTwinForDesiredPropertiesPatch(id);
-                }
-            }
-
-            public async Task Update(string id, Twin twin)
-            {
-                Events.UpdatingTwin(id);
-                await this.twinStore.PutOrUpdate(
-                    id,
-                    new TwinStoreEntity(twin),
-                    twinInfo =>
-                    {
-                        return twinInfo.Twin
-                            .Filter(
-                                storedTwin => storedTwin.Properties?.Desired?.Version < twin.Properties.Desired.Version
-                                    && storedTwin.Properties?.Reported?.Version < twin.Properties.Reported.Version)
-                            .Map(_ => new TwinStoreEntity(Option.Maybe(twin), twinInfo.ReportedPropertiesPatch))
-                            .GetOrElse(twinInfo);
-                    });
-                Events.DoneUpdatingTwin(id);
-            }
-        }
-
-        internal class CloudSync
-        {
-            readonly IConnectionManager connectionManager;
-            readonly IMessageConverter<TwinCollection> twinCollectionConverter;
-            readonly IMessageConverter<Twin> twinConverter;
-
-            public CloudSync(
-                IConnectionManager connectionManager,
-                IMessageConverter<TwinCollection> twinCollectionConverter,
-                IMessageConverter<Twin> twinConverter)
-            {
-                this.connectionManager = connectionManager;
-                this.twinCollectionConverter = twinCollectionConverter;
-                this.twinConverter = twinConverter;
-            }
-
-            public async Task<Option<Twin>> GetTwin(string id)
-            {
-                try
-                {
-                    Events.GettingTwin(id);
-                    Option<ICloudProxy> cloudProxy = await this.connectionManager.GetCloudConnection(id);
-                    Option<Twin> twin = await cloudProxy.Map(
-                            async cp =>
-                            {
-                                IMessage twinMessage = await cp.GetTwinAsync();
-                                Twin twinValue = this.twinConverter.FromMessage(twinMessage);
-                                Events.GetTwinSucceeded(id);
-                                return Option.Some(twinValue);
-                            })
-                        .GetOrElse(() => Task.FromResult(Option.None<Twin>()));
-                    return twin;
-                }
-                catch (Exception ex)
-                {
-                    Events.ErrorGettingTwin(id, ex);
-                    return Option.None<Twin>();
-                }
-            }
-
-            public async Task<bool> UpdateReportedProperties(string id, TwinCollection patch)
-            {
-                try
-                {
-                    Events.UpdatingReportedProperties(id);
-                    Option<ICloudProxy> cloudProxy = await this.connectionManager.GetCloudConnection(id);
-                    bool result = await cloudProxy.Map(
-                            async cp =>
-                            {
-                                IMessage patchMessage = this.twinCollectionConverter.ToMessage(patch);
-                                await cp.UpdateReportedPropertiesAsync(patchMessage);
-                                Events.UpdatedReportedProperties(id);
-                                return true;
-                            })
-                        .GetOrElse(() => Task.FromResult(false));
-                    return result;
-                }
-                catch (Exception ex)
-                {
-                    Events.ErrorUpdatingReportedProperties(id, ex);
-                    return false;
-                }
-            }
-        }
-
         static class Events
         {
             static readonly ILogger Log = Logger.Factory.CreateLogger<StoringTwinManager>();
@@ -466,34 +215,16 @@ namespace Microsoft.Azure.Devices.Edge.Hub.Core.Twin
             {
                 ErrorInDeviceConnectedCallback = IdStart,
                 StoringTwinManagerCreated,
-                DoneSyncingReportedProperties,
                 NoSyncTwinNotStored,
                 UpdatingTwinOnDeviceConnect,
-                ErrorSyncingReportedPropertiesToCloud,
                 SendDesiredPropertyUpdates,
-                StoringReportedPropertiesInStore,
-                UpdatingReportedPropertiesInStore,
-                SyncingReportedPropertiesToCloud,
-                StoredReportedPropertiesFound,
-                UpdateReportedPropertiesSucceeded,
-                UpdateReportedPropertiesFailed,
-                MergedReportedProperties,
-                MergedDesiredProperties,
                 UpdatingDesiredProperties,
-                DoneUpdatingTwin,
-                UpdatingTwin,
-                GettingTwin,
                 GettingTwinFromStore,
                 GotTwinFromTwin,
                 TwinSyncedRecently,
-                GetTwinSucceeded,
-                ErrorGettingTwin,
                 UpdatingReportedProperties,
-                UpdatedReportedProperties,
-                ErrorUpdatingReportedProperties,
                 ErrorHandlingDeviceConnected,
-                HandlingDeviceConnectedCallback,
-                NoTwinForDesiredPropertiesPatch
+                HandlingDeviceConnectedCallback
             }
 
             public static void ErrorInDeviceConnectedCallback(Exception ex)
@@ -516,99 +247,19 @@ namespace Microsoft.Azure.Devices.Edge.Hub.Core.Twin
                 Log.LogWarning((int)EventIds.ErrorHandlingDeviceConnected, ex, $"Error handling device connected event for {id}");
             }
 
-            public static void ErrorUpdatingReportedProperties(string id, Exception ex)
-            {
-                Log.LogDebug((int)EventIds.ErrorUpdatingReportedProperties, ex, $"Error updating reported properties for {id}");
-            }
-
-            public static void UpdatedReportedProperties(string id)
-            {
-                Log.LogInformation((int)EventIds.UpdatedReportedProperties, $"Updated reported properties for {id}");
-            }
-
             public static void UpdatingReportedProperties(string id)
             {
                 Log.LogDebug((int)EventIds.UpdatingReportedProperties, $"Updating reported properties for {id}");
             }
 
-            public static void ErrorGettingTwin(string id, Exception ex)
-            {
-                Log.LogWarning((int)EventIds.ErrorGettingTwin, ex, $"Error getting twin for {id}");
-            }
-
-            public static void GetTwinSucceeded(string id)
-            {
-                Log.LogDebug((int)EventIds.GetTwinSucceeded, $"Got twin for {id}");
-            }
-
-            public static void GettingTwin(string id)
-            {
-                Log.LogDebug((int)EventIds.GettingTwin, $"Getting twin for {id}");
-            }
-
-            public static void DoneUpdatingTwin(string id)
-            {
-                Log.LogDebug((int)EventIds.DoneUpdatingTwin, $"Updated twin in store for {id}");
-            }
-
-            public static void UpdatingTwin(string id)
-            {
-                Log.LogDebug((int)EventIds.UpdatingTwin, $"Updating twin in store for {id}");
-            }
-
-            public static void MergedDesiredProperties(string id)
-            {
-                Log.LogDebug((int)EventIds.MergedDesiredProperties, $"Merged desired properties for {id} in store");
-            }
-
             public static void UpdatingDesiredProperties(string id)
             {
                 Log.LogDebug((int)EventIds.UpdatingDesiredProperties, $"Updating desired properties for {id}");
-            }
-
-            public static void MergedReportedProperties(string id)
-            {
-                Log.LogDebug((int)EventIds.MergedReportedProperties, $"Merged reported properties in store for {id}");
-            }
-
-            public static void UpdateReportedPropertiesFailed(string id)
-            {
-                Log.LogWarning((int)EventIds.UpdateReportedPropertiesFailed, $"Updating reported properties failed {id}");
-            }
-
-            public static void UpdateReportedPropertiesSucceeded(string id)
-            {
-                Log.LogDebug((int)EventIds.UpdateReportedPropertiesSucceeded, $"Updated reported properties for {id}");
-            }
-
-            public static void StoredReportedPropertiesFound(string id)
-            {
-                Log.LogDebug((int)EventIds.StoredReportedPropertiesFound, $"Found stored reported properties for {id} to sync to cloud");
-            }
-
-            public static void SyncingReportedPropertiesToCloud(string id)
-            {
-                Log.LogDebug((int)EventIds.SyncingReportedPropertiesToCloud, $"Syncing stored reported properties to cloud in {id}");
-            }
-
-            public static void UpdatingReportedPropertiesInStore(string id, TwinCollection patch)
-            {
-                Log.LogDebug((int)EventIds.UpdatingReportedPropertiesInStore, $"Updating reported properties in store with version {patch.Version} for {id}");
-            }
-
-            public static void StoringReportedPropertiesInStore(string id, TwinCollection patch)
-            {
-                Log.LogDebug((int)EventIds.StoringReportedPropertiesInStore, $"Storing reported properties in store for {id} with version {patch.Version}");
-            }
+            }            
 
             public static void SendDesiredPropertyUpdates(string id)
             {
                 Log.LogDebug((int)EventIds.SendDesiredPropertyUpdates, $"Sending desired property updates to {id}");
-            }
-
-            public static void ErrorSyncingReportedPropertiesToCloud(string id, Exception ex)
-            {
-                Log.LogDebug((int)EventIds.ErrorSyncingReportedPropertiesToCloud, ex, $"Error syncing reported properties to cloud for {id}");
             }
 
             public static void GettingTwinFromStore(string id)
@@ -634,21 +285,6 @@ namespace Microsoft.Azure.Devices.Edge.Hub.Core.Twin
             public static void UpdatingTwinOnDeviceConnect(string id)
             {
                 Log.LogDebug((int)EventIds.UpdatingTwinOnDeviceConnect, $"Updated twin for {id} on device connect event");
-            }
-
-            public static void DoneSyncingReportedProperties(string id)
-            {
-                Log.LogInformation((int)EventIds.DoneSyncingReportedProperties, $"Done syncing reported properties for {id}");
-            }
-
-            internal static void ErrorSyncingReportedPropertiesToCloud(Exception e)
-            {
-                Log.LogWarning((int)EventIds.ErrorSyncingReportedPropertiesToCloud, e, $"Error in pump to sync reported properties to cloud");
-            }
-
-            public static void NoTwinForDesiredPropertiesPatch(string id)
-            {
-                Log.LogInformation((int)EventIds.NoTwinForDesiredPropertiesPatch, $"Cannot store desired properties patch  for {id} in store as twin was not found");
             }
         }
     }
